Update unit tests for MosseDistribution
